Update nova patch to support qemu commandline in nova for vdpa ports
Adding a new calss LibvirtConfigQemuCommandLine which creates xml
section for qemu:commandline tag.
A qemu section will only added when the port type is virtio-forwarder,
and the alias name of the interface will be generated from the interface
index in net list.
As the normal and macvtap ports will have the same alias format so we
are increasing the index in case of these ports as well